+++ libft.h
@@ -56,6 +56,6 @@ size_t	ft_strlen_s(const char *s);
 ssize_t	ft_putuint_fd(unsigned int nb, int fd);
 int		ft_islower(int c);
 int		ft_isupper(int c);
-ssize_t	ft_putlowerhex_fd(unsigned int n, int fd);
-ssize_t	ft_putupperhex_fd(unsigned int n, int fd);
+ssize_t	ft_putlowerhex_fd(size_t n, int fd);
+ssize_t	ft_putupperhex_fd(size_t n, int fd);
 #endif
